requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# YOLO detection
ultralytics>=8.0.0

# Tracking
filterpy>=1.4.5
scipy>=1.10.0
lap>=0.4.0

# Action recognition models
pytorchvideo>=0.1.5
fvcore>=0.1.5

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0

# Optional: Pose estimation (install separately if needed)
# mmpose>=1.0.0
# mmcv>=2.0.0
# mmengine>=0.7.0
# Installation: pip install mmpose mmcv mmengine (may require additional setup)
